// packages/b_declarations/src/properties/font-variant-caps/types.ts
// b_path:: packages/b_declarations/src/properties/font-variant-caps/types.ts

export type FontVariantCapsIR = {
  kind: "keyword";
  value:
    | "normal"
    | "small-caps"
    | "all-small-caps"
    | "petite-caps"
    | "all-petite-caps"
    | "unicase"
    | "titling-caps"
    | "inherit"
    | "initial"
    | "revert"
    | "revert-layer"
    | "unset";
};